silicon_flow_generator/pyproject.toml
[project]
name = "silicon-flow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "llama-index>=0.14.19",
    "openai>=2.30.0",
    "pydantic-settings>=2.13.1",
    "pymysql>=1.1.2",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "python-json-logger>=4.1.0",
    "python-socketio>=5.16.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
